Preprocess pupil data using this single function

This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path and outputs the preprocessed data to a specified folder path

preprocess(import = NULL, pattern = "*.txt", taskname = NULL,
  subj.prefix = NULL, subj.suffix = NULL, output = NULL,
  gazedata.include = FALSE, eyetracker = NULL, hz = NULL,
  eye.recorded = NULL, eye.use = NULL,
  startrecording.message = "default", startrecording.match = "exact",
  trialonset.message = NULL, trialonset.match = "exact",
  pretrial.duration = NULL, missing.allowed = 1, interpolate = NULL,
  interpolate.maxgap = Inf, smooth = NULL, smooth.window = 5,
  method.first = NULL, bc = NULL, bc.duration = NULL,
  baselineoffset.message = NULL, baselineoffset.match = "exact",
  downsample.binlength = NULL, subset = "default",
  trial.exclude = c())

Arguments

import
Folder path to raw data files

pattern
Pattern to look for in data files

taskname
Name of task - to be used in naming pre-processed files

subj.prefix
The unique pattern prefix (letter(s) and/or symbol(s)) that comes before the subject number in the data file

subj.suffix
The unique pattern suffix (letter(s) or symbol(s)) that comes after the subject number in the data file

output
Folder path to output preprocessed data to

gazedata.include
Logical. Include columns for x and y coordinates of eye gaze? (Default: FALSE)

eyetracker
Which eye-tracker was used to record data

hz
At which frequency was pupil data sampled at? (only required for interpolation and smoothing)

eye.recorded
Do you want to inclue the "left", "right', or "both" eyes?

eye.use
Which eye to use? Left or right

startrecording.message
Message used in SMI experiment to mark StartTracking inline

startrecording.match
Should the message string be an "exact" match or a "pattern" match?

trialonset.message
Message string that marks the start of a trial

trialonset.match
Should the message string be an "exact" match or a "pattern" match?

pretrial.duration
Duration of pre-trial baseline period in milliseconds

missing.allowed
What proportion of missing data is allowed, on a trial-by-trial basis? (Default: 1)

interpolate
What type of interpolation to use? linear or cubic-spline

interpolate.maxgap
Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.

smooth
The type of smoothing function to apply. hann or moving window average (mwa)

smooth.window
Window size of smoothing function default is 5 milliseconds

method.first
Should "smooth" or "interpolate" be applied first? (default: NULL)

bc
Do you want to use "subtractive" or "divisive" baseline correction? (default: "subtractive")

bc.duration
Duration baseline period(s) to use for correction

baselineoffset.message
Message string(s) that marks the offset of baseline period(s)

baselineoffset.match
Message string(s) that marks the offset of baseline period(s)

downsample.binlength
Length of bins to average (default: NULL)

subset
Which columns in the raw data output file do you want to keep

trial.exclude
Specify if ther are any trials to exclude. Trial number

Examples