Convert "messy" raw data file to "tidy" raw data file

This function converts a "messy" raw eye-tracker data file to a "tidy" raw data file.

pupil_read(file, eyetracker = "", starttracking.message = "default",
  starttracking.match = "exact", subj.prefix = NULL,
  subj.suffix = NULL, timing.file = NULL, subset = "default",
  trial.exclude = c())

Arguments

file
A file path to the raw data export file

eyetracker
Which eye-tracker system was used to record data?

starttracking.message
Message used in SMI experiment to mark StartTracking inline

starttracking.match
Should the message string be an "exact" match or a "pattern" match?

subj.prefix
The unique pattern prefix (letter(s) and/or symbol(s)) that comes before the subject number in the data file

subj.suffix
The unique pattern suffix (letter(s) or symbol(s)) that comes after the subject number in the data file

timing.file
File location and name that contains timing information for message markers

subset
Which columns in the raw data export file do you want to keep

trial.exclude
Specify if ther are any trials to exclude. Trial number

Details

Different eye-trackers will save the raw data file using completely different organizations and naming conventions This function will create a "tidy" raw data file that has a standard organization and naming convention, regardless of which eye-tracker was used. Supported Eye-Tracker Systems: Sensomotoric Instruments ("smi") and SR-Research EyeLink ("eyelink")