Deprecated functions in package pupillometry.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

aoi.downsample(x, bin.length = NULL, aoi_col = c())

# S3 method for comparison
plot(import, files = c(), labels = c(), trial,
  title = "", legend.title = "Labels")

pupil.baselinecorrect(x, baselineoffset.message = "",
  bc.duration = 200, bc.type = "subtractive", match = "exact")

pupil.cor(x)

pupil.downsample(x, bin.length = NULL, bc = FALSE)

pupil.eye(x, eye.recorded = "", eye.use = "",
  gazedata.include = FALSE)

pupil.interpolate(x, type = "cubic-spline", maxgap = Inf, hz = "")

pupil.missing(x, missing.allowed = 1, velocity = "", margin = "")

pupil.smooth(x, type = "hann", window = 5, hz = "")

set.stimuli(x)

set.timing(x, trialonset.message = "", ms.conversion = 1,
  pretrial.duration = 0, match = "exact")

set.trial(x, startrecording.message = "", match = "exact")

aoi.downsample

For aoi.downsample, use aoi_downsample.

plot.comparison

For plot.comparison, use plot_comparison.

pupil.baselinecorrect

For pupil.baselinecorrect, use pupil_baselinecorrect.

pupil.cor

For pupil.cor, use pupil_cor.

pupil.downsample

For pupil.downsample, use pupil_downsample.

pupil.eye

For pupil.eye, use pupil_eye.

pupil.interpolate

For pupil.interpolate, use pupil_interpolate.

pupil.missing

For pupil.missing, use pupil_missing.

pupil.smooth

For pupil.smooth, use pupil_smooth.

set.stimuli

For set.stimuli, use set_stimuli.

set.timing

For set.timing, use set_timing.

set.trial

For set.trial, use set_trial.

Examples

aoi.downsample(x, bin.length = 100)

#> Warning: 'aoi.downsample' is deprecated.
#> Use 'aoi_downsample' instead.
#> See help("Deprecated")

#> Error in dplyr::group_by(x, Trial): object 'x' not found

pupil.baselinecorrect(file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")

#> Error in pupil.baselinecorrect(file = "path/filename", baseline.duration = 2000,     start.trial = "# Message: Target"): unused arguments (file = "path/filename", baseline.duration = 2000, start.trial = "# Message: Target")

pupil.cor(x)

#> Warning: 'pupil.cor' is deprecated.
#> Use 'pupil_cor' instead.
#> See help("Deprecated")

#> Error in dplyr::mutate(x, Pupils.r = stats::cor(L_Pupil_Diameter.mm, R_Pupil_Diameter.mm,     use = "pairwise.complete.obs"), Pupils.r = ifelse(is.na(Pupils.r),     0, Pupils.r)): object 'x' not found

pupil.downsample(x, bin.length = 100)

#> Warning: 'pupil.downsample' is deprecated.
#> Use 'pupil_downsample' instead.
#> See help("Deprecated")

#> Error in dplyr::group_by(x, Trial): object 'x' not found

pupil.interpolate(x, type = "cubic-spline", eye.recorded = "both")

#> Error in pupil.interpolate(x, type = "cubic-spline", eye.recorded = "both"): unused argument (eye.recorded = "both")

pupil.missing(x)

#> Warning: 'pupil.missing' is deprecated.
#> Use 'pupil_missing' instead.
#> See help("Deprecated")

#> Error in dplyr::group_by(x, Trial): object 'x' not found

pupil.smooth(x, type = "hann", window = 11, eye.recorded = "both")

#> Error in pupil.smooth(x, type = "hann", window = 11, eye.recorded = "both"): unused argument (eye.recorded = "both")

set.stimuli(x)

#> Warning: 'set.stimuli' is deprecated.
#> Use 'set_stimuli' instead.
#> See help("Deprecated")

#> Error in dplyr::group_by(x, Trial): object 'x' not found

set.timing(x, start.trial = "Fixation", ms.conversion = 1000)

#> Error in set.timing(x, start.trial = "Fixation", ms.conversion = 1000): unused argument (start.trial = "Fixation")

set.trial(x, start.trial = "Fixation", ms.conversion = 1000)

#> Error in set.trial(x, start.trial = "Fixation", ms.conversion = 1000): unused arguments (start.trial = "Fixation", ms.conversion = 1000)