Preprocess pupil data using this single function

This function will perform preprocessing on an entire folder of data files. It takes as input a specified folder path and outputs the preprocessed data to a specified folder path

pupil_preprocess(import.dir = NULL, pattern = "*.txt",
  taskname = NULL, subj.prefix = NULL, subj.suffix = NULL,
  timing.file = NULL, output.dir = NULL, output.steps = FALSE,
  eyetracker = NULL, hz = NULL, eye.use = NULL,
  px_to_mm.conversion = NULL, start_tracking.message = "default",
  start_tracking.match = "exact", trial_onset.message = NULL,
  trial_onset.match = "exact", deblink.extend = 0,
  pre_trial.duration = NULL, missing.allowed = 1, interpolate = NULL,
  interpolate.maxgap = Inf, smooth = NULL, smooth.window = 5,
  method.first = NULL, bc = NULL, pre_bc.duration = NULL,
  bc_onset.message = NULL, bc_onset.match = "exact",
  subset = "default", trial.exclude = c(), files.merge = FALSE,
  starttracking.message = NULL, starttracking.match = NULL,
  trialonset.message = NULL, trialonset.match = NULL,
  pretrial.duration = NULL, prebc.duration = NULL,
  bconset.message = NULL, bconset.match = NULL)

Arguments

import.dir
Folder path to raw data files

pattern
Pattern to look for in data files

taskname
Name of task - to be used in naming pre-processed files

subj.prefix
The unique pattern prefix (letter(s) and/or symbol(s)) that comes before the subject number in the data file

subj.suffix
The unique pattern suffix (letter(s) or symbol(s)) that comes after the subject number in the data file

timing.file
File location and name that contains timing information for message markers

output.dir
Folder path to output preprocessed data to

output.steps
Output files for each step in preprocessing?

eyetracker
Which eye-tracker was used to record data

hz
At which frequency was pupil data sampled at? (only required for interpolation and smoothing)

eye.use
Which eye to use? Left or right

px_to_mm.conversion
The conversion factor to go from px pupil diameter to mm pupil diameter

start_tracking.message
Message used in SMI experiment to mark StartTracking inline

start_tracking.match
Should the message string be an "exact" match or a "pattern" match?

trial_onset.message
Message string that marks the start of a trial

trial_onset.match
Should the message string be an "exact" match or a "pattern" match

deblink.extend
How many milliseconds to extend blinks before and after blink detection

pre_trial.duration
Duration of pre-trial baseline period in milliseconds

missing.allowed
What proportion of missing data is allowed, on a trial-by-trial basis? (Default: 1)

interpolate
What type of interpolation to use? linear or cubic-spline

interpolate.maxgap
Maximum number of NAs to interpolate over. Anything gaps over this value will not be interpolated.

smooth
The type of smoothing function to apply. hann or moving window average (mwa)

smooth.window
Window size of smoothing function default is 5 milliseconds

method.first
Should "smooth" or "interpolate" be applied first? (default: NULL)

bc
Do you want to use "subtractive" or "divisive" baseline correction? (default: "subtractive")

pre_bc.duration
Duration baseline period(s) to use for correction

bc_onset.message
Message string(s) that marks the offset of baseline period(s)

bc_onset.match
Message string(s) that marks the offset of baseline period(s)

subset
Which columns in the raw data output file do you want to keep

trial.exclude
Specify if ther are any trials to exclude. Trial number

files.merge
Do you want to create a single merge output file?

starttracking.message
See start_tracking.message

starttracking.match
See start_tracking.match

trialonset.message
See trial_onset.message

trialonset.match
See trial_onset.match

pretrial.duration
See pre_trial.duration

prebc.duration
See pre_bc.duration

bconset.message
See bc_onset.duration

bconset.match
See bc_onset.match